backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
psycopg2-binary==2.9.10
geojson==3.2.0
requests==2.32.3